Use arc4random_addrandom() only when defined (v2)
Explanation: Android libc is based on OpenBSD libc in which
there not exist anymore arc4random_addrandom().

This patch is inspired by the patch (link below) that was
written by Landry Breuil for Mozilla.

This is the second iteration of the patch, that was simplified
to make it easier to read and understand.

Ref.:
- https://bugzilla.mozilla.org/show_bug.cgi?id=931354
- https://bug931354.bmoattachments.org/attachment.cgi?id=829728